+++ MoreFertilizers/MoreFertilizers/docs/Changelog.md
@@ -7,6 +7,7 @@
 * Tree Tapper's: Automate integration.
 * Something for mushroom boxen? (1.6 - reduce to one day?)
 * Make good draw code for all of the fertilizers. (don't forget domesticated fish - maybe some more jumping...)
+* Everlasting Fruit Fertilizer + Soil Rebalance Fertilizer?
 
 #### Version 0.3.5
 * Buff the bountiful bush fertilizer.